gen_assists_docs skip hidden files
